refactor(common): 순환 의존 제거 + enrichment.py 파사드 분해
Goal A — summarizer↔summary_quality, enrichment↔summary_quality 삼각 순환 제거:
- boilerplate 감지기 3개를 summary_quality로 통합(단방향 provider화)
- deferred # noqa: E402 mid-file import 전부 제거
- enrichment/summarizer는 summary_quality를 top-level 단방향 의존

Goal B — enrichment.py(2110→1186줄) 책임 단위 분해(파사드 유지):
- enrichment_images.py(신규): 이미지 URL 검증
- enrichment_synthetic.py(신규): 합성 설명·제목/엔티티 분석·relevance/dedup
- enrichment.py는 이동한 public 심볼 전부 re-export(기존 import 경로 하위호환)
- test_enrichment.py: @patch 타깃 15개를 새 모듈 네임스페이스로 재배치

네트워크 코어(URL 해석/페이지 파싱)는 mock 결합도가 높아 이번 범위서 제외(경계 명시).

검증: pytest 4753 passed/16 skipped(불변), ruff check/format 클린,
basedpyright 0 errors, 순환 없음(import 정상).
