+++ docs/component-counts.md
@@ -10,7 +10,7 @@
 | --- | --- |
 | 수집기 (scripts/collect_*.py) | 13 |
 | 생성기 (scripts/generate_*.py) | 6 |
-| 공통 모듈 (scripts/common/**/*.py, __init__ 제외) | 55 |
+| 공통 모듈 (scripts/common/**/*.py, __init__ 제외) | 57 |
 | GitHub Actions 워크플로우 (.github/workflows/*.yml) | 48 |
 | 카테고리 페이지 (pages/*.md) | 15 |
 | 테스트 파일 (tests/test_*.py) | 113 |
